+++ breeze-serialization/src/test/kotlin/SerializerTest.kt
@@ -11,6 +11,12 @@ class SerializerTest {
 		val a = mapOf("name" to "Windea", "gender" to "Female")
 		val b = """{"name":"Windea","gender":"Female"}""".trimIndent()
 		assertEquals(a, b.deserializeBy(DataType.Json))
+	}
+
+	@Test
+	fun test2(){
+		val a = mapOf("name" to "Windea", "gender" to "Female")
+		val b = """{"name":"Windea","gender":"Female"}""".trimIndent()
 		assertEquals(b, a.serializeBy<Map<*,*>>(DataType.Json))
 	}
 }